requirements/local.txt
-r base.txt

django-debug-toolbar==2.2
black==19.10b0
flake8==3.8.1
pytest-cov==2.9.0
pytest-django==3.9.0
pytest-pythonpath==0.7.3
pre-commit==2.7.1
django-memcache-status==2.2
